Integrated the update sources dialog into sources list

+++ console-web/webapp/src/main/java/org/openthinclient/web/pkgmngr/ui/presenter/SourcesListPresenter.java
@@ -11,18 +11,29 @@ import com.vaadin.ui.Notification;
 import com.vaadin.ui.Table;
 import com.vaadin.ui.TextArea;
 import com.vaadin.ui.TextField;
+import com.vaadin.ui.UI;
 import com.vaadin.ui.themes.ValoTheme;
 
 import org.openthinclient.pkgmgr.PackageManager;
 import org.openthinclient.pkgmgr.db.Source;
 import org.openthinclient.pkgmgr.db.SourceRepository;
+import org.openthinclient.pkgmgr.op.PackageListUpdateReport;
+import org.openthinclient.pkgmgr.progress.ListenableProgressFuture;
+import org.openthinclient.web.progress.ProgressReceiverDialog;
 import org.openthinclient.web.ui.event.PackageManagerTaskActivatedEvent;
 import org.openthinclient.web.ui.event.PackageManagerTaskFinalizedEvent;
+import org.vaadin.dialogs.ConfirmDialog;
 import org.vaadin.spring.events.Event;
 import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 
+import java.net.MalformedURLException;
+import java.net.URL;
+
 public class SourcesListPresenter {
 
+    public static final String FIELD_URL = "url";
+    public static final String FIELD_DESCRIPTION = "description";
+    public static final String FIELD_ENABLED = "enabled";
     private final View view;
     private final FieldGroup sourceFormBinder;
     private final BeanItemContainer<Source> container;
@@ -32,9 +43,9 @@ public class SourcesListPresenter {
         this.view = view;
 
         sourceFormBinder = new FieldGroup();
-        sourceFormBinder.bind(view.getURLTextField(), "url");
-        sourceFormBinder.bind(view.getDescriptionTextArea(), "description");
-        sourceFormBinder.bind(view.getEnabledCheckBox(), "enabled");
+        sourceFormBinder.bind(view.getURLTextField(), FIELD_URL);
+        sourceFormBinder.bind(view.getDescriptionTextArea(), FIELD_DESCRIPTION);
+        sourceFormBinder.bind(view.getEnabledCheckBox(), FIELD_ENABLED);
 
         view.getURLTextField().setConverter(new StringToUrlConverter());
 
@@ -48,34 +59,52 @@ public class SourcesListPresenter {
         this.view.getAddSourceButton().addClickListener(this::addSourceClicked);
         this.view.getDeleteSourceButton().addClickListener(this::removeSourceClicked);
         this.view.getUpdateButton().addClickListener(this::updateSourcesClicked);
+        this.view.getUpdateButtonTop().addClickListener(this::updateSourcesClicked);
     }
 
     private void updateSourcesClicked(Button.ClickEvent clickEvent) {
-        packageManager.updateCacheDB();
+
+        // deselect any currently selected source
+        sourceSelected(null);
+
+        final ListenableProgressFuture<PackageListUpdateReport> update = packageManager.updateCacheDB();
+
+        final ProgressReceiverDialog dialog = new ProgressReceiverDialog("Updating package database");
+        dialog.watch(update);
+        dialog.open(true);
+
     }
 
     private void removeSourceClicked(Button.ClickEvent clickEvent) {
 
         // FIXME add some kind of confirm dialog
-        Source source = view.getSelectedSource();
-        sourceSelected(null);
-        container.removeItem(source);
 
-        if (source.getId() == null) {
-            // this source has not yet been added to the repository.
-            return;
-        } else {
-            // the source has already been persisted. Remove the entity from the database
-            packageManager.getSourceRepository().delete(source);
-        }
+        ConfirmDialog.show(UI.getCurrent(), "Delete source?", "Are you sure that you would like to delete this source?", "Yes", "No", () ->
+        {
+            Source source = view.getSelectedSource();
+            sourceSelected(null);
+            container.removeItem(source);
 
-        view.refreshSourcesList();
+            if (source.getId() == null) {
+                // this source has not yet been added to the repository.
+                return;
+            } else {
+                // the source has already been persisted. Remove the entity from the database
+                packageManager.getSourceRepository().delete(source);
+            }
 
+            view.refreshSourcesList();
+        });
     }
 
     private void addSourceClicked(Button.ClickEvent clickEvent) {
 
         final Source newSource = new Source();
+        try {
+            newSource.setUrl(new URL("http://archive.openthinclient.org/openthinclient/v2.1/manager-rolling/"));
+        } catch (MalformedURLException e) {
+            // should never happen, as the URL is hardcoded
+        }
         newSource.setEnabled(true);
         newSource.setDescription("Newly created source");
         container.addItem(newSource);
@@ -136,6 +165,10 @@ public class SourcesListPresenter {
             final Item sourceItem = getSourceItem(source);
             sourceFormBinder.setItemDataSource(sourceItem);
 
+            // if the source has been updated (that means, a package list has been downloaded)
+            // no further editing of the URL is allowed
+            view.getURLTextField().setEnabled(source.getLastUpdated() == null);
+
         }
 
     }
@@ -187,6 +220,8 @@ public class SourcesListPresenter {
         CheckBox getEnabledCheckBox();
 
         TextArea getDescriptionTextArea();
+
+        Button getUpdateButtonTop();
 
         Table getSourcesTable();
